fix: All the fixes identified during the PR discussion have been implemented.
