+++ app/[locale]/signin/page.tsx
@@ -1,9 +1,10 @@
-import { SignInForm } from './components/signInForm';
+import { SignInForm } from '@/components/auth/signInForm';
+import { Card } from '@/components/ui/card';
 
 export default function SignIn() {
   return (
-    <>
+    <Card>
       <SignInForm />
-    </>
+    </Card>
   );
 }
