+++ app/[locale]/signup/page.tsx
@@ -1,9 +1,10 @@
-import { SignUpForm } from './components/signUpForm';
+import { Card } from '@/components/ui/card';
+import { SignUpForm } from '@/components/auth/signUpForm';
 
 export default function SignUp() {
   return (
-    <>
+    <Card>
       <SignUpForm />
-    </>
+    </Card>
   );
 }
