+++ next-auth.d.ts
@@ -1,13 +1,13 @@
-import NextAuth, { DefaultSession, DefaultUser } from "next-auth"
+import { DefaultSession, DefaultUser } from 'next-auth';
 
-declare module "next-auth" {
+declare module 'next-auth' {
   interface Session {
     user: {
-      id: string
-    } & DefaultSession["user"]
+      id: string;
+    } & DefaultSession['user'];
   }
 
   interface User extends DefaultUser {
-    id: string
+    id: string;
   }
 }
