+++ package.json
@@ -13,8 +13,11 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^5.2.1",
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-slot": "^1.2.3",
+    "bcrypt": "^6.0.0",
+    "bcryptjs": "^3.0.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.542.0",
