+++ src/components/ui/index.ts
@@ -1,1 +1,10 @@
 export { Button } from './button';
+export {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from './card';
+export { Input } from '@/components/ui/input';
+export { Label } from '@/components/ui/label';
